# storage/api-docs/api-docs-v1.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Book App API",
        "description": "API documentation for the Book App",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API Base URL"
        }
    ],
    "paths": {
...
        "/profile": {
            "put": {
                "tags": [
                    "User Profile"
                ],
                "summary": "Update the authenticated user's profile information",
                "operationId": "a7a2f6c18c018dfff944b57cd0054bd3",
                "requestBody": {
                    "$ref": "#/components/requestBodies/UpdateProfileRequestBody"
                },
                "responses": {
                    "200": {
                        "description": "Profile updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/SuccessResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "properties": {
                                                        "user": {
                                                            "$ref": "#/components/schemas/UserResource"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/profile/picture": {
            "put": {
                "tags": [
                    "User Profile"
                ],
                "summary": "Update the authenticated user's profile picture",
                "operationId": "bc24794be20749ec13bdd43a204ca16c",
                "requestBody": {
                    "$ref": "#/components/requestBodies/UpdatePictureRequestBody"
                },
                "responses": {
                    "200": {
                        "description": "Profile picture updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/SuccessResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "properties": {
                                                        "picture_url": {
                                                            "type": "string",
                                                            "example": "/storage/user/profile_pictures/1.jpg"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "No picture uploaded",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "SuccessResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Operation successful"
                    },
                    "data": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "An error occurred"
                    },
                    "errors": {
                        "type": "object",
                        "nullable": true,
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
            "UserResource": {
                "description": "Represents a user in the system",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "example": "johndoe@example.com"
                    },
                    "picture": {
                        "type": "string",
                        "example": "/storage/user/profile_pictures/1.jpg",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00Z"
                    }
                },
                "type": "object"
            }
        },
        "requestBodies": {
...
            "UpdatePictureRequestBody": {
                "required": true,
                "content": {
                    "multipart/form-data": {
                        "schema": {
                            "required": [
                                "picture"
                            ],
                            "properties": {
                                "picture": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "UpdateProfileRequestBody": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "name",
                                "email"
                            ],
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "example": "John Doe Updated"
                                },
                                "email": {
                                    "type": "string",
                                    "example": "newemail@example.com"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        }
    },
...
}